chore: move Error class to infrastructure
